pyproject.toml
[tool.poetry]
name = "MeshSegNet-DentalArchAligner"
version = "0.1.0"
description = ""
authors = ["0x11111111 <dypro_mike@outlook.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
charset-normalizer = "2.1.1"
contourpy = "1.0.6"
cycler = "0.11.0"
cython = "0.29.32"
urllib3 = "1.26.15"
deprecated = "1.2.13"
fonttools = "4.38.0"
idna = "3.4"
joblib = "1.2.0"
jsonpatch = "1.32"
jsonpointer = "2.3"
kiwisolver = "1.4.4"
matplotlib = "3.6.2"
networkx = "2.8.8"
numpy = "1.26.0"
nvidia-cublas-cu11 = "11.10.3.66"
nvidia-cuda-nvrtc-cu11 = "11.7.99"
nvidia-cuda-runtime-cu11 = "11.7.99"
nvidia-cudnn-cu11 = "^9.1.1.17"
packaging = "23.1"
pandas = "1.5.2"
pillow = "9.4.0"
pip = "22.3.1"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
pytz = "2022.7"
requests = "2.32.3"
scikit-learn = "1.2.0"
scipy = "1.13.0"
setuptools = "65.5.0"
six = "1.16.0"
threadpoolctl = "3.1.0"
torch = "2.0.1"
torchaudio = "0.13.1"
torchvision = "0.15.2"
tornado = "6.2"
typing-extensions = "4.6.1"
tzdata = { url = "https://github.com/python/tzdata/archive/refs/tags/2022.7.tar.gz" }
vedo = "^2024.5.2"
visdom = "0.2.3"
vtk = "9.2.4"
websocket-client = "1.4.2"
wheel = "0.37.1"
wrapt = "1.14.1"
onnx = "1.14.0"
onnxruntime = "1.15.1"
pycocotools = "2.0.7"
PyYAML = "6.0.1"
onnxsim = "0.4.36"
onnxruntime-gpu = "1.18.0"
opencv-python = "4.9.0.80"
psutil = "5.9.8"
py-cpuinfo = "9.0.0"
huggingface-hub = "0.23.2"
safetensors = "0.4.3"
meshlib = "^2.4.2.30"
pyunpack = "^0.3"
nibabel = "^5.2.1"
